* gcc.c-torture/execute/longlong.c (alpha_ep_extbl_i_eq_0): Allow 	for 16-bit 'unsigned int'.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@45304 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.c-torture/execute/longlong.c
@@ -10,7 +10,7 @@ void alpha_ep_extbl_i_eq_0()
 {
   unsigned int rb, ra, rc;
 
-  rb  = (((unsigned int)(pars) >> 27)) & 0x1fUL;
+  rb  = (((unsigned long)(pars) >> 27)) & 0x1fUL;
   ra  = (((unsigned int)(pars) >> 5)) & 0x1fUL;
   rc  = (((unsigned int)(pars) >> 0)) & 0x1fUL;
   {
